completely overhauled the naming conventions of swot_bot for better clarity, easier readability and to negate the need for namespaced routes. removed _profile from every instance of student_profile, teacher_profile and guardian_profile in the entire program. This includes renaming files. Also removed all user_ from user_identities.
All model tests currently pass and the seeds file is populating the database correctly as well.

In this instance, we dropped the database to save ourselves a significant amount time. We know that once this app is deployed, we will not be able to do this and it is considered poor practice, but in this instance with all the changes to be made, we felt it was necessary.