Security fix: require guardianship to view student
This controller was authenticating users but not authorizing them. Via
Guardians::BaseController, this controller required the user to be a guardian
(authenticating them as a real user), but didn't check that they have a
relationship to the student (authorizing them to perform this action).

Instead of looking up the student by id directly, this changes it to look up
the student through Guardian's relationship to Student. This find() will only
search the guardian's students.

There are similar problems all over this codebase. I wondered about it after
your presentation this morning mentioned the importance of security. Using
relationships is the easiest way to scope searches and prevent people from
looking up others' records. Use this spec as an example to test your other
controllers.