Fix the bug introduced in #258
The bug that was occuring here was a result of assignment_categories
with no assignments in them. In order to remedy this, we updated
GradeCalculator to specify a starting value for the map over each
category (re: `reduce(0, :+)`). This was the exact same bug that occurred
in #253 that was subsequently fixed in #256. (Note to selves, if using
reduce with an array that may be empty, __specify the starting value__).

As we were rounding grade all over the place, `Student#grade_in` now
rounds the grade to 1 decimal point automatically.

`students/courses/show` was using the legacy
`Student#calculate_percent`. Updated that.

To finish off the commit, we added a test that makes sure the
GradeCalculator can handle the error condition that existed before
(e.g. empty courses/assignment_categories). This involved adding an
identity to the Student factory.