Add attendance screen.
Working toward fixing rwarbelow/the_swot_bot#217.

This commit makes attendance a first-class citizen in the application
(meaning it adds a table to track it). Currently, the only way
to access attendance is through StudentAction (and StudentActionType and
StudentActionCategory). That would equate to approximately 3 queries,
per student, per day in the report. This solution gathers the same data
with one query per student.

This particular commit also does a few infrastructure things:

* Add CSV as a Mime-Type so you can specify it as a Rails output (e.g.
  attendance.csv.erb)

* Add .rb as a Template type so you can use plain old Ruby code to
  generate your view instead of ERB (e.g. attendance.csv.rb)