Finishes rwarbelow/the_swot_bot#217.
Added jQuery-UI to add the datepicker functionality to the Attendance
report in lieu of the pagination that was there a few commits ago.

This additionally fixes a bug we found in the attendance creation where
the test for Attendance through the live-class view wasn't verifying
that the Attendances were for actual students, which caused attendance
to appear incorrectly in the view.

There is also a new migration that populates the Attendance table by
searching all the StudentActions for Attendance related actions and
creating records for each one.

We also refactored the student_attendance method in
Teachers::BaseController to use the Attendance table, which sped it up
by a couple hundred milliseconds.